binance: breakout into `feed` and `broker` mods like other backends
